requirements.txt
anthropic>=0.49.0
psd-tools>=1.9.0
photoshop-python-api>=0.22.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
python-dotenv>=1.0.0
Pillow>=10.0.0
